Moved loginusingSession to AuthCheck and added a basic profile page
